Update README.md
The code consists of 10 classes wherein 5 classes are for international locations and 5 classes are for domestic locations. Other than the locations, there is a class for booking flights where the user will be given options to select his destination. Whereas there is one class where flight status is displayed(whether the flight is on time or delayed. The user will also have an option to web check in and get the boarding pass in advance. There is one extra functionality for hotel booking pre-hand to avoid inconvenience after reaching the destination.For this functionality, file handling was used. Next, while booking for the flight, the user will also have an option to select his food preference well in advance before he boards the flight.
And in the very end, a receipt will be given to the user where all the data given by the user will be displayed along with the cost.